small move from SeekerApplication to NetworkState object
